// src/app/core/service/model/agency.model.ts
export class AgencyModel{
  address: string;
  districtCode: string
  email?: string
  id: string
  managementDeparment: string
  managementUnitCode: string
  managementUnitName: string
  managementUnitStatus: number;
  managementUnitType: number;
  parent: string
  phoneNumber?: string
  provinceCode: string
  wardCode: string
  web: null
}